requirements.txt
telethon
python-dotenv
requests
streamlit
streamlit-autorefresh
pandas
pymongo[srv]
numpy